蓝黄不一致

+++ engine/plugin/plugins/ffmpeg_adapter/muxer/ffmpeg_muxer_plugin.cpp
@@ -371,7 +371,8 @@ Status FFmpegMuxerPlugin::Prepare()
         SetTagsOfTrack(outputFormat_.get(), formatContext_->streams[pair.first], pair.second);
     }
     SetTagsOfGeneral(formatContext_.get(), generalParameters_);
-    formatContext_->flags = (int)((unsigned int)formatContext_->flags | (unsigned int)AVFMT_TS_NONSTRICT);
+    formatContext_->flags = static_cast<unsigned int>(formatContext_->flags) | 
+        static_cast<unsigned int>(AVFMT_TS_NONSTRICT);
     return Status::OK;
 }
 void FFmpegMuxerPlugin::ResetIoCtx(IOContext& ioContext)
